requirements.txt
six>=1.11.0
scikit-learn==0.22.2
numpy>=1.14.5
scipy>=1.5.4
pyreadr>=0.2.8
numba>=0.49.0
tensorflow_gpu>=2.3.1
tensorflow_probability==0.11.1
statsmodels>=0.12.1
scikit-multilearn>=0.2.0
torch>=1.7.1
matplotlib>=3.3.3
seaborn>=0.11.1
h5py>=2.10.0